requirements.txt
beautifulsoup4==4.12.2
cvzone==1.5.6
Flask==2.2.5
gTTS==2.3.2
numpy==1.21.6
openai==0.27.8
opencv_python==4.8.0.74
opencv_python_headless==4.8.0.76
python-dotenv==1.0.0
pyzbar==0.1.9
Requests==2.31.0
ultralytics==8.0.26
Werkzeug==2.2.3
youtube_transcript_api==0.6.1
